refactor(lib/dom): change naming
- Create single file tabbableHelpers.ts
- Delete all unused helpers
- Create getTabbableElements()
